+++ sponj/common/src/main/java/com/chimericdream/sponj/SponjMod.java
@@ -2,6 +2,7 @@ package com.chimericdream.sponj;
 
 import com.chimericdream.lib.registries.ModRegistryHelper;
 import com.chimericdream.sponj.blocks.ModBlocks;
+import com.chimericdream.sponj.registry.ModStats;
 import com.google.common.base.Suppliers;
 import dev.architectury.registry.registries.RegistrarManager;
 import org.apache.logging.log4j.LogManager;
@@ -17,6 +18,8 @@ public final class SponjMod {
 
     public static void init() {
         MANAGER = Suppliers.memoize(() -> RegistrarManager.get(ModInfo.MOD_ID));
+
+        ModStats.init();
 
         REGISTRY_HELPER.init();
         ModBlocks.init();